Fish10 legacy implementation code refactor
