+++ fish_11_dll/src/dll_interface/fish_10/fish10_decryptmsg.rs
@@ -7,6 +7,7 @@ use crate::dll_interface::utility;
 use crate::platform_types::{BOOL, HWND};
 use crate::unified_error::DllError;
 use crate::{buffer_utils, dll_function_identifier, legacy, log_debug};
+use crate::crypto::blowfish;
 use std::ffi::c_char;
 use std::os::raw::c_int;
 
@@ -55,7 +56,7 @@ fn fish10_decrypt_msg_impl(input_str: &str) -> Result<String, DllError> {
     log_debug!("FiSH10: decrypting message for '{}' with legacy key", target);
 
     // Decrypt using legacy Blowfish algorithm
-    let decrypted = legacy::blowfish::decrypt_message(key, encrypted_message, target.as_bytes())
+    let decrypted = blowfish::decrypt_message(key, encrypted_message, target.as_bytes())
         .map_err(|e| DllError::LegacyError {
             context: format!("Blowfish decryption failed for '{}'", target),
             cause: e.to_string(),
