+++ fish_11_dll/src/dll_interface/fish_10/fish10_encryptmsg.rs
@@ -7,6 +7,7 @@ use crate::dll_interface::utility;
 use crate::platform_types::{BOOL, HWND};
 use crate::unified_error::DllError;
 use crate::{buffer_utils, dll_function_identifier, legacy, log_debug};
+use crate::crypto::blowfish;
 use std::ffi::c_char;
 use std::os::raw::c_int;
 
@@ -37,7 +38,7 @@ fn fish10_encrypt_msg_impl(input: &str) -> Result<String, DllError> {
     log_debug!("FiSH10: Encrypting message for '{}' with legacy key", target);
 
     // Encrypt using legacy Blowfish algorithm
-    let encrypted = legacy::blowfish::encrypt_message(key, plaintext_message, target.as_bytes())
+    let encrypted = blowfish::encrypt_message(key, plaintext_message, target.as_bytes())
         .map_err(|e| DllError::LegacyError {
             context: format!("Blowfish encryption failed for '{}'", target),
             cause: e.to_string(),
