+++ fish_11_dll/src/dll_interface/fish_10/mod.rs
@@ -4,7 +4,6 @@
 
 pub mod fish10_decryptmsg;
 pub mod fish10_delkey;
-pub mod fish10_dh1080;
 pub mod fish10_encryptmsg;
 pub mod fish10_gettopicsetting;
 pub mod fish10_haskey;
